Move saving logic into results component
