+++ src/Components/SavedPage/SavedProject/SavedProject.css
@@ -9,6 +9,7 @@
   border: 1px solid #DBD5D5;
   background: #FFF;
   box-shadow: 0px 1px 1px 0px rgba(0, 0, 0, 0.25);
+  margin: 20px;
 }
 
 .saved-project-title, .saved-project-description {
