--- src/types.ts
@@ -1,15 +1,0 @@
-export type Project = {
-    id: string,
-    type: string,
-    attributes: {
-    name: string,
-    steps: string,
-    description: string,
-    features: string,
-    interactions: string,
-    colors: string,
-    timeline: string,
-    saved: boolean,
-    user_id: number
-    }
-}
